Database Update
Anytime we ran the /scores route, the request would time out and not work.

Turns out somehow we weren't connecting to the database, and that was causing the suspension of the request being made.

Updated server.js to reflect database connection changes.

Also added CORS, specifying localhost:8080, as well as the website URL Kim wanted whitelisted.