User endpoints implemented.
Most of the user endpoints are created.

Set it up in server.js where if the user is authenticated (have an account), to update the last login when they use an API endpoint in the database if that record exists in the user collection. Still finishing up the logic to create that user into the database if that user is not located.